Added tests for set command based on ScalaTest
