+++ Makefile
@@ -1,0 +1,4 @@
+test:
+	docker compose up -d
+	sbt run test
+	docker compose down
