+++ docker-compose.yml
@@ -2,8 +2,9 @@ version: "3"
 
 services:
   tile38:
-    image: tile38/tile38:1.24.0
-    container_name: tile38
+    image: tile38/tile38:1.30.2
     command: tile38-server -vv -p 9851
+    volumes:
+      - ./src/test/data:/data
     ports:
-      - 9851:9851
+      - "9851:9851"
