+++ project/Dependencies.scala
@@ -1,10 +1,12 @@
 import sbt._
 
 object Dependencies {
-  val lettuce        = "io.lettuce"     % "lettuce-core" % "5.0.2.RELEASE"
+  val lettuce = "io.lettuce" % "lettuce-core" % "5.0.2.RELEASE"
+
+  val scalatest: Seq[ModuleID] = Seq("org.scalatest" %% "scalatest" % "3.2.15" % "test")
 
   val circeVersion = "0.14.1"
-  val circe = Seq(
+  val circe: Seq[ModuleID] = Seq(
     "io.circe" %% "circe-core",
     "io.circe" %% "circe-generic",
     "io.circe" %% "circe-parser",
@@ -12,17 +14,15 @@ object Dependencies {
     "io.circe" %% "circe-shapes"
   ).map(_ % circeVersion)
 
-  val catsVersion = "2.7.0"
-  val cats = Seq(
+  val catsVersion = "2.9.0"
+  val cats: Seq[ModuleID] = Seq(
     "org.typelevel" %% "cats-core" % catsVersion
   )
 
-  val catsEffectVersion = "3.3.12"
-  val catsEffect = Seq(
+  val catsEffectVersion = "3.4.5"
+  val catsEffect: Seq[ModuleID] = Seq(
     "org.typelevel" %% "cats-effect" % catsEffectVersion
   )
 
-  val weaver = Seq("com.disneystreaming" %% "weaver-cats" % "0.7.12" % Test)
-
-  val projectDeps = circe ++ cats ++ catsEffect ++ Seq(lettuce) ++ weaver
+  val projectDeps: Seq[sbt.ModuleID] = circe ++ cats ++ catsEffect ++ Seq(lettuce) ++ scalatest
 }
