+++ project/plugins.sbt
@@ -1,5 +1,4 @@
-addSbtPlugin("com.tapad"         % "sbt-docker-compose" % "1.0.34")
-addSbtPlugin("se.marcuslonnberg" % "sbt-docker"         % "1.5.0")
-addSbtPlugin("org.xerial.sbt"    % "sbt-sonatype"       % "3.9.6")
-addSbtPlugin("com.jsuereth"      % "sbt-pgp"            % "1.1.1")
-addCompilerPlugin("com.olegpy"  %% "better-monadic-for" % "0.3.1")
+addSbtPlugin("org.xerial.sbt"   % "sbt-sonatype"       % "3.9.6")
+addSbtPlugin("com.jsuereth"     % "sbt-pgp"            % "1.1.1")
+addCompilerPlugin("com.olegpy" %% "better-monadic-for" % "0.3.1")
+addSbtPlugin("org.scalameta"    % "sbt-scalafmt"       % "2.5.0")
